fix : mv RecruitRole TeamJob , move directory and following refactor
